sets up tx helpers in the contract utils lib
