--- apps/summon/src/types/form.ts
@@ -1,30 +1,0 @@
-export type FormValues = {
-  daoName?: string;
-  tokenName?: string;
-  tokenSymbol?: string;
-  lootTokenName?: string;
-  lootTokenSymbol?: string;
-  votingTransferable?: boolean;
-  nvTransferable?: boolean;
-  quorum?: string;
-  minRetention?: string;
-  sponsorThreshold?: string;
-  newOffering?: string;
-  votingPeriod?: string;
-  votingPeriodInSeconds?: number;
-  gracePeriod?: string;
-  gracePeriodInSeconds?: number;
-  shamans?:
-    | ''
-    | {
-        shamanAddresses: string[];
-        shamanPermissions: string[];
-      };
-  members?:
-    | ''
-    | {
-        memberAddresses: string[];
-        memberShares: string[];
-        memberLoot: string[];
-      };
-};
