+++ apps/summon/src/utils/summonlegos.ts
@@ -6,7 +6,7 @@ const BaalSummonerContract: ContractLego = {
   contractName: 'BaalSummoner',
   type: 'static',
   abi: LOCAL_ABI.BAAL_ADV_TOKEN_SUMMONER,
-  targetAddress: CONTRACT_KEYCHAINS.V3_FACTORY,
+  targetAddress: CONTRACT_KEYCHAINS.V3_FACTORY_ADV,
 };
 
 export const SummonTX = {
