+++ libs/contract-utils/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/contract-meta';
 export * from './lib/moloch-v3-contract';
 export * from './lib/moloch-v3-summoner-contract';
+export * from './lib/summon-tx-utils';
 export * from './lib/encoding-utils';
 export * from './lib/estimate-util';
 export * from './lib/types';
