+++ libs/contract-utils/src/lib/summon-tx-utils.ts
@@ -1,6 +1,7 @@
 import { LOCAL_ABI } from '@daohaus/abis';
 import {
   ArgType,
+  DAOHAUS_SUMMONER_REFERRER,
   encodeFunction,
   encodeValues,
   getNonce,
@@ -14,10 +15,38 @@ import {
 } from '@daohaus/utils';
 import { CONTRACT_KEYCHAINS, ValidNetwork } from '@daohaus/keychain-utils';
 
-import { FormValues } from '../types/form';
-import { FORM_KEYS } from './formKeys';
+export type SummonParams = {
+  daoName?: string;
+  tokenName?: string;
+  tokenSymbol?: string;
+  lootTokenName?: string;
+  lootTokenSymbol?: string;
+  votingTransferable?: boolean;
+  nvTransferable?: boolean;
+  quorum?: string;
+  minRetention?: string;
+  sponsorThreshold?: string;
+  newOffering?: string;
+  votingPeriod?: string;
+  votingPeriodInSeconds?: number;
+  gracePeriod?: string;
+  gracePeriodInSeconds?: number;
+  shamans?:
+    | ''
+    | {
+        shamanAddresses: string[];
+        shamanPermissions: string[];
+      };
+  members?:
+    | ''
+    | {
+        memberAddresses: string[];
+        memberShares: string[];
+        memberLoot: string[];
+      };
+};
 
-export const encodeMintParams = (formValues: FormValues) => {
+export const encodeMintParams = (formValues: SummonParams) => {
   const { members } = formValues;
 
   if (
@@ -53,7 +82,7 @@ export const encodeMintParams = (formValues: FormValues) => {
   throw new Error('Encoding Error');
 };
 
-export const encodeTokenParams = (formValues: FormValues) => {
+export const encodeTokenParams = (formValues: SummonParams) => {
   const tokenName = formValues.tokenName;
   const tokenSymbol = formValues.tokenSymbol;
   const lootTokenName = formValues.lootTokenName;
@@ -94,7 +123,7 @@ export const encodeTokenParams = (formValues: FormValues) => {
   throw new Error('Encoding Error');
 };
 
-const governanceConfigTX = (formValues: FormValues) => {
+const governanceConfigTX = (formValues: SummonParams) => {
   const {
     votingPeriodInSeconds,
     gracePeriodInSeconds,
@@ -137,7 +166,7 @@ const governanceConfigTX = (formValues: FormValues) => {
   throw new Error('Encoding Error');
 };
 
-export const shamanConfigTX = (formValues: FormValues) => {
+export const shamanConfigTX = (formValues: SummonParams) => {
   const { shamans } = formValues;
 
   if (shamans === '' || !shamans) {
@@ -168,7 +197,7 @@ export const shamanConfigTX = (formValues: FormValues) => {
   throw new Error('Encoding Error');
 };
 
-const metadataConfigTX = (formValues: FormValues, posterAddress: string) => {
+const metadataConfigTX = (formValues: SummonParams, posterAddress: string) => {
   const { daoName } = formValues;
   if (!isString(daoName)) {
     console.log('ERROR: Form Values', formValues);
@@ -191,14 +220,92 @@ const metadataConfigTX = (formValues: FormValues, posterAddress: string) => {
   throw new Error('Encoding Error');
 };
 
-const handleKeychains = (chainId: ValidNetwork) => {
-  const { V3_FACTORY, POSTER } = CONTRACT_KEYCHAINS;
-  const v3Contracts = [V3_FACTORY, POSTER];
+// THIS IS ONLY USED IN V3_FACTORY_ORIGINAL CONTRACT
+const tokenConfigTX = (formValues: SummonParams) => {
+  const pauseVoteToken = !formValues.votingTransferable;
+  const pauseNvToken = !formValues.nvTransferable;
+
+  const encoded = encodeFunction(LOCAL_ABI.BAAL, 'setAdminConfig', [
+    pauseVoteToken,
+    pauseNvToken,
+  ]);
+
+  if (isString(encoded)) {
+    return encoded;
+  }
+  throw new Error('Encoding Error');
+};
+
+// THIS IS ONLY USED IN V3_FACTORY_ORIGINAL CONTRACT
+export const shareConfigTX = (formValues: SummonParams) => {
+  const { members } = formValues;
+
+  if (
+    !members ||
+    !isArray(members?.memberAddresses) ||
+    members.memberAddresses.some((addr) => !isString(addr)) ||
+    !isArray(members?.memberShares) ||
+    members.memberShares.some((shares) => !isNumberish(shares))
+  ) {
+    console.log('ERROR: Form Values', formValues);
+    throw new Error(
+      'shareConfigTX recieved arguments in the wrong shape or type'
+    );
+  }
+
+  const wholeShareAmts = members.memberShares;
+  const sharesInBaseUnits = wholeShareAmts.map((shares) => toBaseUnits(shares));
+  const encoded = encodeFunction(LOCAL_ABI.BAAL, 'mintShares', [
+    members.memberAddresses,
+    sharesInBaseUnits,
+  ]);
+
+  if (isString(encoded)) {
+    return encoded;
+  }
+  throw new Error('Encoding Error');
+};
+
+// THIS IS ONLY USED IN V3_FACTORY_ORIGINAL CONTRACT
+export const lootConfigTX = (formValues: SummonParams) => {
+  const { members } = formValues;
+
+  if (
+    !members ||
+    !isArray(members?.memberAddresses) ||
+    members.memberAddresses.some((addr) => !isString(addr)) ||
+    !isArray(members?.memberShares) ||
+    members.memberShares.some((shares) => !isNumberish(shares))
+  ) {
+    console.log('ERROR: Form Values', formValues);
+    throw new Error(
+      'shareConfigTX recieved arguments in the wrong shape or type'
+    );
+  }
+
+  const wholeLootAmts = members.memberLoot;
+  const lootInBaseUnits = wholeLootAmts.map((loot) =>
+    toBaseUnits(loot.toString())
+  );
+  const encoded = encodeFunction(LOCAL_ABI.BAAL, 'mintLoot', [
+    members.memberAddresses,
+    lootInBaseUnits,
+  ]);
+  if (isString(encoded)) {
+    return encoded;
+  }
+  throw new Error('Encoding Error');
+};
+
+export const handleKeychains = (chainId: ValidNetwork) => {
+  const { V3_FACTORY_ADV, POSTER, ZODIAC_FACTORY } = CONTRACT_KEYCHAINS;
+  const v3Contracts = [V3_FACTORY_ADV, POSTER];
 
   if (v3Contracts.every((contract) => contract[chainId])) {
     return {
-      V3_FACTORY: V3_FACTORY[chainId] || '',
+      V3_FACTORY_ADV: V3_FACTORY_ADV[chainId] || '',
       POSTER: POSTER[chainId] || '',
+      ZODIAC_FACTORY: ZODIAC_FACTORY[chainId] || '',
     };
   }
 
@@ -207,12 +314,13 @@ const handleKeychains = (chainId: ValidNetwork) => {
 
 export const assembleTxArgs = (
   formValues: Record<string, unknown>,
-  chainId: ValidNetwork
+  chainId: ValidNetwork,
+  safeAddress?: string
 ): ArgType[] => {
-  const tokenName = formValues[FORM_KEYS.TOKEN_NAME];
-  const tokenSymbol = formValues[FORM_KEYS.TOKEN_SYMBOL];
-  const lootTokenName = formValues[FORM_KEYS.LOOT_TOKEN_NAME];
-  const lootTokenSymbol = formValues[FORM_KEYS.LOOT_TOKEN_SYMBOL];
+  const tokenName = formValues['tokenName'];
+  const tokenSymbol = formValues['tokenSymbol'];
+  const lootTokenName = formValues['lootTokenName'];
+  const lootTokenSymbol = formValues['lootTokenSymbol'];
 
   if (
     !isString(tokenName) ||
@@ -239,13 +347,58 @@ export const assembleTxArgs = (
     metadataConfigTX(formValues, POSTER),
   ];
   const args = [
-    ZERO_ADDRESS,
+    safeAddress || ZERO_ADDRESS,
     ZERO_ADDRESS,
     getNonce(),
     mintParams,
     tokenParams,
     initActions,
   ];
+  console.log('args', args);
+
+  return args;
+};
+
+// THIS IS ONLY USED FOR THE V3_FACTORY_ORIGINAL CONTRACT
+export const assembleTxArgsOriginalFactory = (
+  formValues: Record<string, unknown>,
+  chainId: ValidNetwork,
+  safeAddress?: string
+): ArgType[] => {
+  const tokenName = formValues['tokenName'];
+  const tokenSymbol = formValues['tokenSymbol'];
+
+  if (!isString(tokenName) || !isString(tokenSymbol)) {
+    console.log('ERROR: Form Values', formValues);
+
+    throw new Error(
+      'assembleSummonTx recieved arguments in the wrong shape or type'
+    );
+  }
+
+  const { POSTER } = handleKeychains(chainId);
+
+  const initParams = encodeValues(
+    ['string', 'string', 'address', 'address', 'address', 'address'],
+    [
+      tokenName,
+      tokenSymbol,
+      safeAddress || ZERO_ADDRESS,
+      ZERO_ADDRESS,
+      ZERO_ADDRESS,
+      ZERO_ADDRESS,
+    ]
+  );
+
+  const initActions = [
+    tokenConfigTX(formValues),
+    governanceConfigTX(formValues),
+    shamanConfigTX(formValues),
+    shareConfigTX(formValues),
+    lootConfigTX(formValues),
+    metadataConfigTX(formValues, POSTER),
+  ];
+  const args = [initParams, initActions, getNonce(), DAOHAUS_SUMMONER_REFERRER];
   console.log('args', args);
 
   return args;
