+++ libs/keychain-utils/src/contractKeychains.ts
@@ -1,15 +1,15 @@
 import { Keychain } from './types';
 
 export const CONTRACT_KEYCHAINS: Record<string, Keychain> = {
-  V3_FACTORY: {
-    '0x1': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
-    '0x5': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
-    '0x64': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
-    '0x89': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
-    '0xa': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
-    '0xa4b1': '0x7734b4b08c66086BAC419c9cADe64b0A0491D781',
+  V3_FACTORY_ADV: {
+    '0x1': '0x0',
+    '0x5': '0xb0c5c96c3d21c1d58B98a5366dF0Af7AfcD94F95',
+    '0x64': '0x0',
+    '0x89': '0x0',
+    '0xa': '0x0',
+    '0xa4b1': '0x0',
   },
-  V3_FACTORY_HARDCODED_LOOT: {
+  V3_FACTORY_ORIGINAL: {
     '0x1': '0x7e988A9db2F8597735fc68D21060Daed948a3e8C',
     '0x5': '0x7e988A9db2F8597735fc68D21060Daed948a3e8C',
     '0x64': '0x7e988A9db2F8597735fc68D21060Daed948a3e8C',
