+++ libs/utils/src/constants/contracts.ts
@@ -1,8 +1,8 @@
 import { ValidNetwork } from '@daohaus/keychain-utils';
 
 export type ContractKey =
-  | 'V3_FACTORY_HARDCODED_LOOT'
-  | 'V3_FACTORY'
+  | 'V3_FACTORY_ORIGINAL'
+  | 'V3_FACTORY_ADV'
   | 'LOOT_SINGLETON'
   | 'SHARES_SINGLETON'
   | 'BAAL_SINGLETON'
